Update ruby.md by adding array method
Array#append is more idiomatic in modern Ruby (post-2.5) for its clarity and descriptive name, although Array#push is still frequently used due to its long history and prevalence in older codebases. 

It would be a good practice for Ruby beginners to use append for new code. 

Also, this would more intuitively help them remember Array#prepend for adding to the start of an array (as opposed to Array#unshift)